Home work to lesson 6

+++ wnteredshop/api/src/main/java/ru/geekbrains/wnteredshop/api/CartDto.java
@@ -1,5 +1,6 @@
 package ru.geekbrains.wnteredshop.api;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
@@ -11,14 +12,14 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 }
