+++ wnteredshop/core-service/src/main/resources/db/migration/V2__orders.sql
@@ -3,7 +3,7 @@ CREATE TABLE orders(
                 username VARCHAR(255)  ,
                 address VARCHAR(255),
                 phone VARCHAR(255),
-                total_price int,
+                total_price DECIMAL,
                 created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                 updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
@@ -12,8 +12,8 @@ CREATE TABLE order_items(
                 product_id bigint REFERENCES products(id),
                 order_id bigint REFERENCES orders(id),
                 quantity int,
-                price_per_product int,
-                price int,
+                price_per_product DECIMAL,
+                price DECIMAL,
                 created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                 updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
